Split the logic for rdb migration and trials cache
