+++ optuna_dashboard/_trials.py
@@ -1,0 +1,52 @@
+from __future__ import annotations
+
+from datetime import datetime
+from datetime import timedelta
+import threading
+
+from optuna.storages import BaseStorage
+from optuna.storages import RDBStorage
+from optuna.trial import FrozenTrial
+from optuna.version import __version__ as optuna_ver
+from packaging import version
+
+
+# In-memory trials cache
+trials_cache_lock = threading.Lock()
+trials_cache: dict[int, list[FrozenTrial]] = {}
+trials_last_fetched_at: dict[int, datetime] = {}
+
+
+def get_trials(storage: BaseStorage, study_id: int) -> list[FrozenTrial]:
+    with trials_cache_lock:
+        trials = trials_cache.get(study_id, None)
+
+        # Not a big fan of the heuristic, but I can't think of anything better.
+        if trials is None or len(trials) < 100:
+            ttl_seconds = 2
+        elif len(trials) < 500:
+            ttl_seconds = 5
+        else:
+            ttl_seconds = 10
+
+        last_fetched_at = trials_last_fetched_at.get(study_id, None)
+        if (
+            trials is not None
+            and last_fetched_at is not None
+            and datetime.now() - last_fetched_at < timedelta(seconds=ttl_seconds)
+        ):
+            return trials
+    trials = storage.get_all_trials(study_id, deepcopy=False)
+
+    if (
+        # See https://github.com/optuna/optuna/pull/3702
+        version.parse(optuna_ver) <= version.Version("3.0.0rc0.dev")
+        and isinstance(storage, RDBStorage)
+        and storage.url.startswith("postgresql")
+    ):
+        trials = sorted(trials, key=lambda t: t.number)
+
+    with trials_cache_lock:
+        trials_last_fetched_at[study_id] = datetime.now()
+        trials_cache[study_id] = trials
+    return trials
